refactor: dto 패키지 수정
