+++ src/test/java/chess/dao/FakeGameDao.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-import chess.dao.dto.GameDto;
+import chess.dto.GameDto;
 import chess.domain.GameStatus;
 
 import java.util.ArrayList;
